% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionCuad_Pers.R
\name{CuadradoPearson}
\alias{CuadradoPearson}
\title{Cuadrado de Pearson}
\usage{
CuadradoPearson(
  datos,
  req,
  kilos = 1000,
  p1 = 1,
  p2 = 2,
  e1 = 1,
  e2 = 2,
  precio1 = 40,
  precio2 = 70,
  precio3 = 80,
  precio4 = 70
)
}
\arguments{
\item{datos}{de la variable respuesta}

\item{req}{es la necesidad que necesitan de proteina que necesitan nuestros animales.}

\item{kilos}{es la cantidad de kilos que queremos producir de alimento.}

\item{p1}{es el alimento proteico 1}

\item{p2}{es el alimento proteico 2}

\item{e1}{es el alimento energético 1}

\item{e2}{es el alimento energético 2}

\item{precio1}{es el precio 1 del alimento proteico 1}

\item{precio2}{es el precio del alimento proteico 2}

\item{precio3}{es el precio del alimento enegético 1}

\item{precio4}{es le precio del alimento energético 2}
}
\value{
devuelve una tabla donde esta $Tabla_MayorReq la cual nos indica los alimentos proteicos que son mayores al requerimiento que necesitamos y tambien la cantidad de proteina que contiene cada alimento,tambien una tabla llamada $Tabla_MenorReq la cual no indica los alimentos energéticos que son menores al requerimiento,posteriormente tenemos una tabla en donde sólo estan los dos alimentos de la clasificacion 1 y la clasificación 2 que cumplen con las condiciones que nos indica el cuadrado de Pearson y juento con ellas no da el porcentaje de proteina que contiene cada una de ellas ,finalmente nos devuelve una tabla final en donde tenemos el prrocentaje de proteina bruta de cada alimento,las diferencias,una comprobación de si es correctyo o noel aporte de proteína de cada uno de los alimentos,y el precio total por alimento y la suma de los precios de todos los alimentos,de igual forma elabora un gráfico de pastel  de proteinas donde se observa el aporte de proteína de cada uno de los alimentos,y otro gráfico en histograma  donde se observa el precio de cada alimento y ver cual de estos alimentos es más caro.
}
\description{
El modelo Doble Cuadrado de Pearson permite balancear raciones,la cual tiene limitaciones no se pueden utilizar muchos ingredientes.Consiste en realizar un cuadro donde en el extremo superior,se pone el valor de proteína que aportan los dos primeros alimentos proteicos los cuales deben de estar por encima del requerimiento y en la esquina inferior se ponen los dos alimentos energéticos los cuales estarán por debajo del requerimiento y en el centro se pone el valor del requerimientos que necesitamos.
}
